+++ src/mailman/runners/bounce.py
@@ -51,12 +51,17 @@ class BounceRunner(Runner):
         addresses = StandardVERP().get_verp(mlist, msg)
         if len(addresses) > 0:
             # Scan the message to see if it contained permanent or temporary
-            # failures.  We'll ignore temporary failures, but even if there
-            # are no permanent failures, we'll assume VERP bounces are
-            # permanent.
+            # failures.  We'll ignore temporary failures, and if there
+            # are no permanent failures, we'll assume this is a vacation
+            # response or similar.
             temporary, permanent = all_failures(msg)
             if len(temporary) > 0:
                 # This was a temporary failure, so just ignore it.
+                return False
+            if len(permanent) == 0:
+                log.info('VERPed bounce message but not a recognized DSN: %s',
+                         msg.get('message-id', 'n/a'))
+                maybe_forward(mlist, msg)
                 return False
         else:
             # See if this was a probe message.