+++ src/mailman/docs/NEWS.rst
@@ -104,6 +104,8 @@ Bugs
   without specifying policy.  (Closes #934)
 * Improve error message from REST API for invalid email address (Fixes #872)
 * Nonmember posts gated from usenet bypass nonmember moderation.  (Closes #937)
+* VERPed vacation responses and similar are no longer scored as bounces.
+  (Closes #939)
 
 Command line
 ------------